general updates done to the landing page, and cleaner code implimented
