--- src/invoice-gen/hooks/index.ts
@@ -1,1 +1,0 @@
-export const nothing = "nothing";
