+++ src/landing/components/FeaturesSection.tsx
@@ -1,5 +1,4 @@
 import { motion } from "framer-motion";
-
 import features from "../constants/features";
 
 const containerVariants = {
