+++ src/landing/components/Navbar.tsx
@@ -1,24 +1,23 @@
-import { useState } from "react";
 import { Button } from "@/shared/components/ui/button";
-import { Link } from "react-router-dom";
-import { Menu, X } from "lucide-react";
+import { Link, useLocation } from "react-router-dom";
+import { Menu, X, ChevronDown } from "lucide-react";
+import { useNavbar } from "../hooks/navhooks";
+import { featureLinks } from "../constants/navbar";
 
 const Navbar = () => {
-  const [isMenuOpen, setIsMenuOpen] = useState(false);
-  const toggleMenu = () => setIsMenuOpen(!isMenuOpen);
+  const {
+    isMenuOpen,
+    setIsMenuOpen,
+    toggleMenu,
+  } = useNavbar();
 
-  const scrollToTop = (e: React.MouseEvent<HTMLAnchorElement>) => {
-    e.preventDefault();
-    window.scrollTo({ top: 0, behavior: "smooth" });
-  };
+  const location = useLocation();
+  const currentPath = location.pathname;
 
-  const scrollToSection = (id: string) => (e: React.MouseEvent<HTMLAnchorElement>) => {
-    e.preventDefault();
-    const section = document.querySelector(id);
-    if (section) {
-      section.scrollIntoView({ behavior: "smooth" });
-    }
-  };
+  const isHomeActive = currentPath === "/";
+  const isProgramsActive = currentPath === "/programs";
+  const isBlogsActive = currentPath === "/blogs";
+  const isPricingActive = currentPath === "/pricing";
 
   return (
     <nav className="absolute top-6 left-0 right-0 mx-20 bg-[#03192F] rounded-2xl py-4 z-50 shadow-xl">
@@ -34,35 +33,65 @@ const Navbar = () => {
 
           <div className="hidden md:block">
             <div className="ml-10 flex items-center space-x-4">
-              <a
-                href="#"
-                onClick={scrollToTop}
-                className="px-3 py-2 text-sm font-medium text-white transition-colors"
+              <Link
+                to="/"
+                className={`px-3 py-2 text-sm font-medium transition-colors ${
+                  isHomeActive
+                    ? "text-indigo-400 font-semibold border-b-2 border-indigo-500"
+                    : "text-white"
+                }`}
               >
                 Home
-              </a>
-              <a
-                href="#features"
-                onClick={scrollToSection("#features")}
-                className="px-3 py-2 text-sm font-medium text-white transition-colors"
-              >
-                Features
-              </a>
+              </Link>
+
+              <div className="relative group">
+                <div className="px-3 py-2 text-sm font-medium text-white flex items-center gap-1 cursor-pointer">
+                  Features <ChevronDown size={16} />
+                </div>
+
+                <div className="absolute top-10 left-0 w-64 bg-white rounded-lg shadow-lg z-50 p-2 flex flex-col gap-1 opacity-0 invisible group-hover:opacity-100 group-hover:visible transition-all duration-200">
+                  {featureLinks.map(({ path, label, icon: Icon }, index) => (
+                    <Link
+                      key={index}
+                      to={path}
+                      className="flex items-center gap-3 px-4 py-2 text-sm text-gray-800 rounded-md hover:bg-gray-100 hover:text-indigo-600 transition-colors"
+                    >
+                      <Icon size={24} />
+                      {label}
+                    </Link>
+                  ))}
+                </div>
+              </div>
+
               <Link
-                to="/pricing"
-                className="px-3 py-2 text-sm font-medium text-white transition-colors"
+                to="/programs"
+                className={`px-3 py-2 text-sm font-medium transition-colors ${
+                  isProgramsActive
+                    ? "text-indigo-400 font-semibold border-b-2 border-indigo-500"
+                    : "text-white"
+                }`}
               >
                 Programs
               </Link>
+
               <Link
-                to="/about"
-                className="px-3 py-2 text-sm font-medium text-white transition-colors"
+                to="/blogs"
+                className={`px-3 py-2 text-sm font-medium transition-colors ${
+                  isBlogsActive
+                    ? "text-indigo-400 font-semibold border-b-2 border-indigo-500"
+                    : "text-white"
+                }`}
               >
                 Blogs
               </Link>
+
               <Link
                 to="/pricing"
-                className="px-3 py-2 text-sm font-medium text-white transition-colors"
+                className={`px-3 py-2 text-sm font-medium transition-colors ${
+                  isPricingActive
+                    ? "text-indigo-400 font-semibold border-b-2 border-indigo-500"
+                    : "text-white"
+                }`}
               >
                 Pricing
               </Link>
@@ -100,52 +129,66 @@ const Navbar = () => {
         </div>
       </div>
 
-      {/* Mobile menu */}
+      {/* Mobile Menu */}
       {isMenuOpen && (
         <div className="md:hidden neo-blur border-b border-white/10">
           <div className="px-2 pt-2 pb-3 space-y-1 sm:px-3">
-            <a
-              href="#"
-              onClick={(e) => {
-                scrollToTop(e);
-                toggleMenu();
-              }}
-              className="block px-3 py-2 rounded-md text-base font-medium text-white"
+            <Link
+              to="/"
+              onClick={toggleMenu}
+              className={`block px-3 py-2 rounded-md text-base font-medium ${
+                isHomeActive ? "text-indigo-400 font-semibold" : "text-white"
+              }`}
             >
               Home
-            </a>
+            </Link>
+
             <a
               href="#features"
               onClick={(e) => {
-                scrollToSection("#features")(e);
+                e.preventDefault();
+                const section = document.querySelector("#features");
+                if (section) {
+                  section.scrollIntoView({ behavior: "smooth" });
+                }
                 toggleMenu();
               }}
               className="block px-3 py-2 rounded-md text-base font-medium text-white"
             >
               Features
             </a>
+
             <Link
-              to="/pricing"
+              to="/programs"
               onClick={toggleMenu}
-              className="block px-3 py-2 rounded-md text-base font-medium text-white"
+              className={`block px-3 py-2 rounded-md text-base font-medium ${
+                isProgramsActive ? "text-indigo-400 font-semibold" : "text-white"
+              }`}
             >
               Programs
             </Link>
+
             <Link
-              to="/about"
+              to="/blogs"
               onClick={toggleMenu}
-              className="block px-3 py-2 rounded-md text-base font-medium text-white"
+              className={`block px-3 py-2 rounded-md text-base font-medium ${
+                isBlogsActive ? "text-indigo-400 font-semibold" : "text-white"
+              }`}
             >
               Blogs
             </Link>
+
             <Link
               to="/pricing"
               onClick={toggleMenu}
-              className="block px-3 py-2 rounded-md text-base font-medium text-white"
+              className={`block px-3 py-2 rounded-md text-base font-medium ${
+                isPricingActive ? "text-indigo-400 font-semibold" : "text-white"
+              }`}
             >
               Pricing
             </Link>
           </div>
+
           <div className="pt-4 pb-3 border-t border-white/10">
             <div className="flex items-center px-5">
               <div className="flex-shrink-0 w-full">
