+++ src/landing/hooks/navhooks.ts
@@ -1,0 +1,32 @@
+import { useRef, useState, useEffect } from "react";
+
+export const useNavbar = () => {
+  const [isMenuOpen, setIsMenuOpen] = useState(false);
+  const [isDropdownOpen, setIsDropdownOpen] = useState(false);
+  const dropdownRef = useRef<HTMLDivElement>(null);
+
+  const toggleMenu = () => setIsMenuOpen((prev) => !prev);
+
+  useEffect(() => {
+    const handleClickOutside = (event: MouseEvent) => {
+      if (
+        dropdownRef.current &&
+        !dropdownRef.current.contains(event.target as Node)
+      ) {
+        setIsDropdownOpen(false);
+      }
+    };
+
+    document.addEventListener("mousedown", handleClickOutside);
+    return () => document.removeEventListener("mousedown", handleClickOutside);
+  }, []);
+
+  return {
+    isMenuOpen,
+    setIsMenuOpen,
+    isDropdownOpen,
+    setIsDropdownOpen,
+    toggleMenu,
+    dropdownRef,
+  };
+};
